fixed migration files order issue and added hasColumn guard to

+++ database/migrations/2026_03_10_070946_add_holiday_ot_hours_to_overtime_table.php
@@ -11,9 +11,11 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('over_times', function (Blueprint $table) {
-            $table->integer('holiday_ot_hours')->default(0);
-        });
+	if(!Schema::hasColumn('over_times', 'holiday_ot_hours')) {
+        	Schema::table('over_times', function (Blueprint $table) {
+         	   	$table->integer('holiday_ot_hours')->default(0);
+        	});
+	}
     }
 
     /**
@@ -21,8 +23,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('over_times', function (Blueprint $table) {
-            $table->dropColumn('holiday_ot_hours');
-        });
+	if(Schema::hasColumn('over_times', 'holiday_ot_hours')) {
+        	Schema::table('over_times', function (Blueprint $table) {
+            		$table->dropColumn('holiday_ot_hours');
+        	});
+	}
     }
 };
